backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
supabase==1.0.4
python-dotenv==1.0.0
pytest==7.4.3
httpx>=0.24.0,<0.25.0
python-jose==3.3.0
